requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
django-filter>=24.0
Pillow>=11.0.0
reportlab>=4.2.0
PyYAML>=6.0.2
python-dateutil>=2.9.0
face-recognition>=1.3.0
opencv-python>=4.10.0
numpy>=1.26.0,<2.0
setuptools>=69.0,<70
